Fixes #287 compilation on macOS with clang 14 fails
Note, review the corrections, you may want to edit some of them
(eg. I changed the result type of:
    int set_dyn_env(int sym, int val);
    int add_dyn_env(int sym, int val);
but you may want to change the type of the fields where
they're assigned instead.)

Also, in opengl.lsp, on macOS, we'd have to use frameworks instead of libraries.
Unfortunately, I've found to way to test for the OS in eisl, so
I commented out two lines for linux and added the two lines for macOS.
You'll want to add a test to key both alternatives.

Some unused variables have been removed, and unused parameters
have been declared unused with __attribute__((unused)), but you
may want to use a macro for this instead.

I kept the:
     static const char __attribute__((unused)) rcsid[] = "$";
lines, but with git they've become useless.
In the Makefile, you may use the command:
    git rev-parse HEAD
to get the current git commit ID, and insert a single occurence of it
in the binary.  You may do it in a C variable, or using linker options
to add a specific segment to the binary. See LDFLAGS in makefile.
Then you may retrieve it with:

$ otool -s .data .gitCommitID eisl
eisl:
Contents of (.data,.gitCommitID) section
0000000130b14000	32 31 35 33 31 31 39 39 34 37 31 35 32 39 33 35
0000000130b14010	38 66 38 61 34 36 33 65 66 65 37 61 65 63 34 35
0000000130b14020	35 64 33 63 30 35 37 63

$ otool -s .data .gitCommitID eisl|sed -e s'/^[0-9a-f]*	//'|xxd -r -p
21531199471529358f8a463efe7aec455d3c057c

+++ main.c
@@ -518,7 +518,7 @@ void set_stok_buf(int index, char c)
 
 void replace_stok_buf(char *str)
 {
-    if (strlen(str) > stok.bufsize) {
+    if (strlen(str) > (size_t)stok.bufsize) {
 	int new_bufsize = strlen(str);
 	RESIZE(stok.buf, sizeof(char) * new_bufsize);
 	stok.bufsize = new_bufsize;
